TS: src/components/Entry/DomainArchitectures
